processor/processor-ocs/requirements.txt
kafka-python
python-swiftclient
keystoneauth1 
